# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent System",
    "description": "AI20K Agent System API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/data-model": {
      "get": {
        "tags": [
          "Data Models"
        ],
        "summary": "Get Current Data Model",
        "description": "Lấy DBML và revision hiện tại của dự án.",
        "operationId": "getDataModel",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Data Models"
        ],
        "summary": "Update Current Data Model",
        "description": "Lưu trực tiếp DBML do người dùng chỉnh sửa thủ công.",
        "operationId": "updateDataModel",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/data-model-changes/pending": {
      "get": {
        "tags": [
          "Data Model Changes"
        ],
        "summary": "Get Pending Change Proposal",
        "operationId": "getPendingProjectDataModelChange",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_ChangeProposalDetailResponse__NoneType__"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/data-model-changes/{change_id}": {
      "get": {
        "tags": [
          "Data Model Changes"
        ],
        "summary": "Xem chi tiết một đề xuất thay đổi mô hình dữ liệu",
        "description": "Trả về DBML đề xuất kèm DBML hiện hành để dựng khung so sánh khác biệt.",
        "operationId": "getProjectDataModelChange",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID dự án chứa Data Model",
              "title": "Project Id"
            },
            "description": "ID dự án chứa Data Model"
          },
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID đề xuất thay đổi Data Model",
              "title": "Change Id"
            },
            "description": "ID đề xuất thay đổi Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChangeProposalDetailResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/data-model-changes/{change_id}/accept": {
      "post": {
        "tags": [
          "Data Model Changes"
        ],
        "summary": "Chấp nhận đề xuất và áp dụng vào mô hình dữ liệu hiện tại",
        "description": "Áp dụng DBML đề xuất vào mô hình dữ liệu, tăng revision và trả về mô hình mới.",
        "operationId": "acceptChangeProposal",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID đề xuất thay đổi Data Model",
              "title": "Change Id"
            },
            "description": "ID đề xuất thay đổi Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataModelResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/data-model-changes/{change_id}/reject": {
      "post": {
        "tags": [
          "Data Model Changes"
        ],
        "summary": "Từ chối một đề xuất thay đổi mô hình dữ liệu",
        "description": "Đánh dấu đề xuất là REJECTED; nội dung và revision của mô hình giữ nguyên.",
        "operationId": "rejectChangeProposal",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID đề xuất thay đổi Data Model",
              "title": "Change Id"
            },
            "description": "ID đề xuất thay đổi Data Model"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChangeProposalSummaryResponse_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Code",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông báo lỗi"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Mã lỗi ổn định cho client"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Danh sách chi tiết lỗi theo field"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "error_code"
        ],
        "title": "ApiErrorResponse",
        "description": "Error envelope được Global Exception Handler trả về."
      },
...
      "ApiResponse_ChangeProposalDetailResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangeProposalDetailResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[ChangeProposalDetailResponse]"
      },
      "ApiResponse_ChangeProposalSummaryResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangeProposalSummaryResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[ChangeProposalSummaryResponse]"
      },
...
      "ApiResponse_DataModelResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataModelResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[DataModelResponse]"
      },
...
      "ApiResponse_Union_ChangeProposalDetailResponse__NoneType__": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "description": "Trạng thái phản hồi (luôn là 'success' đối với 2xx)",
            "default": "success"
          },
          "code": {
            "type": "integer",
            "maximum": 299.0,
            "minimum": 200.0,
            "title": "Code",
            "description": "HTTP Status Code (mặc định 200)",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Thông điệp phản hồi cho người dùng",
            "default": "Xử lý thành công"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangeProposalDetailResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dữ liệu payload chính trả về"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApiResponse[Union[ChangeProposalDetailResponse, NoneType]]"
      },
...
      "ChangeProposalDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Định danh đề xuất thay đổi"
          },
          "data_model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Data Model Id",
            "description": "Định danh Data Model liên quan"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Định danh người tạo đề xuất"
          },
          "base_revision": {
            "type": "integer",
            "title": "Base Revision",
            "description": "Revision Data Model khi tạo đề xuất"
          },
          "status": {
            "$ref": "#/components/schemas/DataModelChangeStatus",
            "description": "Trạng thái đề xuất"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Thời điểm khởi tạo"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Thời điểm cập nhật gần nhất"
          },
          "proposed_dbml": {
            "type": "string",
            "title": "Proposed Dbml",
            "description": "DBML do Agent đề xuất"
          },
          "current_dbml": {
            "type": "string",
            "title": "Current Dbml",
            "description": "DBML chính thức đang áp dụng"
          },
          "current_revision": {
            "type": "integer",
            "title": "Current Revision",
            "description": "Revision hiện tại của Data Model"
          },
          "is_outdated": {
            "type": "boolean",
            "title": "Is Outdated",
            "description": "Proposal không còn khớp revision hiện tại"
          }
        },
        "type": "object",
        "required": [
          "id",
          "data_model_id",
          "user_id",
          "base_revision",
          "status",
          "created_at",
          "updated_at",
          "proposed_dbml",
          "current_dbml",
          "current_revision",
          "is_outdated"
        ],
        "title": "ChangeProposalDetailResponse",
        "description": "Chi tiết proposal kèm snapshot hiện hành."
      },
      "ChangeProposalSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Định danh đề xuất thay đổi"
          },
          "data_model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Data Model Id",
            "description": "Định danh Data Model liên quan"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Định danh người tạo đề xuất"
          },
          "base_revision": {
            "type": "integer",
            "title": "Base Revision",
            "description": "Revision Data Model khi tạo đề xuất"
          },
          "status": {
            "$ref": "#/components/schemas/DataModelChangeStatus",
            "description": "Trạng thái đề xuất"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Thời điểm khởi tạo"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Thời điểm cập nhật gần nhất"
          }
        },
        "type": "object",
        "required": [
          "id",
          "data_model_id",
          "user_id",
          "base_revision",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ChangeProposalSummaryResponse",
        "description": "Thông tin tóm tắt một đề xuất thay đổi."
      },
...
      "DataModelChangeStatus": {
        "type": "string",
        "enum": [
          "PROPOSED",
          "ACCEPTED",
          "REJECTED",
          "CONFLICTED"
        ],
        "title": "DataModelChangeStatus",
        "description": "Trạng thái đề xuất thay đổi mô hình dữ liệu (Data Model Change Status)."
      },
...
      "DataModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "ID của Data Model"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "ID dự án sở hữu Data Model"
          },
          "dbml": {
            "type": "string",
            "minLength": 1,
            "title": "Dbml",
            "description": "Snapshot DBML hiện tại"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "description": "Revision phục vụ optimistic locking"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Thời điểm tạo theo ISO 8601"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Thời điểm cập nhật theo ISO 8601"
          },
          "is_outdated": {
            "type": "boolean",
            "title": "Is Outdated",
            "description": "Model không khớp analysis revisions hiện tại"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "dbml",
          "revision",
          "created_at",
          "updated_at",
          "is_outdated"
        ],
        "title": "DataModelResponse",
        "description": "Snapshot Data Model trả về cho Frontend editor."
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Đường dẫn field không hợp lệ"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mô tả lỗi validation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Chi tiết validation của một trường request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDataModelRequest": {
        "properties": {
          "data_model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Data Model Id",
            "description": "ID của Data Model cần cập nhật"
          },
          "dbml": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Dbml",
            "description": "Toàn bộ snapshot DBML mới"
          },
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision",
            "description": "Revision mà client đã tải trước khi chỉnh sửa"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data_model_id",
          "dbml",
          "base_revision"
        ],
        "title": "UpdateDataModelRequest",
        "description": "Payload lưu trực tiếp DBML với revision gốc của client."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}